Internal: Refactor templates manager [ED-18647]
