+++ includes/template-library/manager.php
@@ -3,10 +3,6 @@ namespace Elementor\TemplateLibrary;
 
 use Elementor\Api;
 use Elementor\Core\Common\Modules\Ajax\Module as Ajax;
-use Elementor\Core\Isolation\Wordpress_Adapter;
-use Elementor\Core\Isolation\Wordpress_Adapter_Interface;
-use Elementor\Core\Isolation\Elementor_Adapter;
-use Elementor\Core\Isolation\Elementor_Adapter_Interface;
 use Elementor\Core\Settings\Manager as SettingsManager;
 use Elementor\Includes\TemplateLibrary\Data\Controller;
 use Elementor\TemplateLibrary\Classes\Import_Images;
@@ -51,16 +47,6 @@ class Manager {
 	private $_import_images = null; // phpcs:ignore PSR2.Classes.PropertyDeclaration.Underscore
 
 	/**
-	 * @var Wordpress_Adapter_Interface
-	 */
-	protected $wordpress_adapter = null;
-
-	/**
-	 * @var Elementor_Adapter_Interface
-	 */
-	protected $elementor_adapter = null;
-
-	/**
 	 * Template library manager constructor.
 	 *
 	 * Initializing the template library manager by registering default template
@@ -102,14 +88,6 @@ class Manager {
 		}
 
 		return $this->_import_images;
-	}
-
-	public function set_wordpress_adapter( Wordpress_Adapter_Interface $wordpress_adapter ) {
-		$this->wordpress_adapter = $wordpress_adapter;
-	}
-
-	public function set_elementor_adapter( Elementor_Adapter_Interface $elementor_adapter ): void {
-		$this->elementor_adapter = $elementor_adapter;
 	}
 
 	/**
@@ -357,13 +335,13 @@ class Manager {
 			return $source->move_template_to_folder( $args );
 		}
 
-		if ( 'local' === $args['from_source'] ) {
-			$args = $this->format_args_for_single_action_from_local_to_cloud( $args );
+		$from_source = $this->get_source( $args['from_source'] );
+
+		if ( ! $from_source ) {
+			return new \WP_Error( 'template_error', 'Template source not found.' );
 		}
 
-		if ( 'cloud' === $args['from_source'] ) {
-			$args = $this->format_args_for_single_action_from_cloud_to_local( $args );
-		}
+		$args = $from_source->format_args_for_single_action( $args );
 
 		$template_id = $source->save_item( $args );
 
@@ -387,13 +365,13 @@ class Manager {
 			return new \WP_Error( 'template_error', 'Template source not found.' );
 		}
 
-		if ( 'local' === $args['from_source'] ) {
-			$args = $this->format_args_for_single_action_from_local_to_cloud( $args );
+		$from_source = $this->get_source( $args['from_source'] );
+
+		if ( ! $from_source ) {
+			return new \WP_Error( 'template_error', 'Template source not found.' );
 		}
 
-		if ( 'cloud' === $args['from_source'] ) {
-			$args = $this->format_args_for_single_action_from_cloud_to_local( $args );
-		}
+		$args = $from_source->format_args_for_single_action( $args );
 
 		$template_id = $source->save_item( $args );
 
@@ -406,51 +384,6 @@ class Manager {
 
 	private function is_action_to_same_source( $args ) {
 		return $args['source'] === $args['from_source'];
-	}
-
-	private function format_args_for_single_action_from_local_to_cloud( $args ) {
-		if ( ! $this->is_allowed_to_read_template( [
-			'source' => $args['from_source'],
-			'template_id' => $args['from_template_id'],
-		] ) ) {
-			return new \WP_Error(
-				'template_error',
-				esc_html__( 'You do not have permission to access this template.', 'elementor' )
-			);
-		}
-
-		$document = Plugin::$instance->documents->get( $args['from_template_id'] );
-
-		if ( ! $document ) {
-			return new \WP_Error( 'template_error', 'Document not found.' );
-		}
-
-		$args['content'] = $document->get_elements_data();
-
-		$page = SettingsManager::get_settings_managers( 'page' )->get_model( $args['from_template_id'] );
-		$args['page_settings'] = $page->get_data( 'settings' );
-
-		return $args;
-	}
-
-	private function format_args_for_single_action_from_cloud_to_local( $args ) {
-		$from_source = $this->get_source( $args['from_source'] );
-
-		if ( ! $from_source ) {
-			return new \WP_Error( 'template_error', 'Template source not found.' );
-		}
-
-		$data = $from_source->get_item( $args['from_template_id'] );
-
-		if ( is_wp_error( $data ) || empty( $data['content'] ) ) {
-			return new \WP_Error( 'template_error', 'Unable to format template args.' );
-		}
-
-		$decoded_data = json_decode( $data['content'], true );
-		$args['content'] = $decoded_data['content'];
-		$args['page_settings'] = $decoded_data['page_settings'];
-
-		return $args;
 	}
 
 	/**
@@ -555,7 +488,13 @@ class Manager {
 			return $validate_args;
 		}
 
-		if ( ! $this->is_allowed_to_read_template( $args ) ) {
+		$source = $this->get_source( $args['source'] );
+
+		if ( ! $source ) {
+			return new \WP_Error( 'template_error', 'Template source not found.' );
+		}
+
+		if ( method_exists( $source, 'is_allowed_to_read_template' ) && ! $source->is_allowed_to_read_template( $args ) ) {
 			return new \WP_Error(
 				'template_error',
 				esc_html__( 'You do not have permission to access this template.', 'elementor' )
@@ -564,12 +503,6 @@ class Manager {
 
 		if ( isset( $args['edit_mode'] ) ) {
 			Plugin::$instance->editor->set_edit_mode( $args['edit_mode'] );
-		}
-
-		$source = $this->get_source( $args['source'] );
-
-		if ( ! $source ) {
-			return new \WP_Error( 'template_error', 'Template source not found.' );
 		}
 
 		do_action( 'elementor/template-library/before_get_source_data', $args, $source );
@@ -1096,44 +1029,6 @@ class Manager {
 		return true;
 	}
 
-	private function is_allowed_to_read_template( array $args ): bool {
-		if ( 'remote' === $args['source'] || 'cloud' === $args['source'] ) {
-			return true;
-		}
-
-		if ( null === $this->wordpress_adapter ) {
-			$this->set_wordpress_adapter( new WordPress_Adapter() );
-		}
-
-		if ( ! $this->should_check_permissions( $args ) ) {
-			return true;
-		}
-
-		$post_id = intval( $args['template_id'] );
-		$post_status = $this->wordpress_adapter->get_post_status( $post_id );
-		$is_private_or_non_published = ( 'private' === $post_status && ! $this->wordpress_adapter->current_user_can( 'read_private_posts', $post_id ) ) || ( 'publish' !== $post_status );
-
-		$can_read_template = $is_private_or_non_published || $this->wordpress_adapter->current_user_can( 'edit_post', $post_id );
-
-		return apply_filters( 'elementor/template-library/is_allowed_to_read_template', $can_read_template, $args );
-	}
-
-	private function should_check_permissions( array $args ): bool {
-		if ( null === $this->elementor_adapter ) {
-			$this->set_elementor_adapter( new Elementor_Adapter() );
-		}
-
-		// TODO: Remove $isWidgetTemplate in 3.28.0 as there is a Pro dependency
-		$check_permissions = isset( $args['check_permissions'] ) && false === $args['check_permissions'];
-		$is_widget_template = 'widget' === $this->elementor_adapter->get_template_type( $args['template_id'] );
-
-		if ( $check_permissions || $is_widget_template ) {
-			return false;
-		}
-
-		return true;
-	}
-
 	public function bulk_move_templates( array $args ) {
 		$validate_args = $this->ensure_args( [ 'source', 'from_source', 'from_template_id' ], $args );
 
@@ -1157,9 +1052,13 @@ class Manager {
 			return $source->move_bulk_templates_to_folder( $args );
 		}
 
-		$bulk_args = 'local' === $args['from_source']
-			? $this->format_args_for_bulk_action_from_local( $args )
-			: $this->format_args_for_bulk_action_from_cloud( $args );
+		$from_source = $this->get_source( $args['from_source'] );
+
+		if ( ! $from_source ) {
+			return new \WP_Error( 'template_error', 'Template source not found.' );
+		}
+
+		$bulk_args = $from_source->format_args_for_bulk_action( $args );
 
 		if ( $source->supports_quota() && ! $this->is_action_to_same_source( $args ) ) {
 			$is_quota_valid = $source->validate_quota( $bulk_args );
@@ -1185,66 +1084,6 @@ class Manager {
 		return $bulk_save;
 	}
 
-	private function format_args_for_bulk_action_from_local( $args ) {
-		$bulk_args = [];
-
-		foreach ( $args['from_template_id'] as $from_template_id ) {
-			if ( ! $this->is_allowed_to_read_template( [
-				'source' => $args['from_source'],
-				'template_id' => $from_template_id,
-			] ) ) {
-				continue;
-			}
-
-			$document = Plugin::$instance->documents->get( $from_template_id );
-
-			if ( ! $document ) {
-				continue;
-			}
-
-			$page = SettingsManager::get_settings_managers( 'page' )->get_model( $from_template_id );
-
-			$bulk_args[] = array_merge(
-				$args,
-				[
-					'title' => $document->get_post()->post_title,
-					'type' => $document::get_type(),
-					'content' => $document->get_elements_data(),
-					'page_settings' => $page->get_data( 'settings' ),
-				]
-			);
-		}
-
-		return $bulk_args;
-	}
-
-	private function format_args_for_bulk_action_from_cloud( $args ) {
-		$from_source = $this->get_source( $args['from_source'] );
-
-		if ( ! $from_source ) {
-			return new \WP_Error( 'template_error', 'Template source not found.' );
-		}
-
-		$templates = $from_source->get_bulk_items( $args );
-		$bulk_args = [];
-
-		foreach ( $templates as $template ) {
-			$content = json_decode( $template['content'], true );
-
-			$bulk_args[] = array_merge(
-				$args,
-				[
-					'title' => $template['title'],
-					'type' => $template['type'],
-					'content' => $content['content'],
-					'page_settings' => $content['page_settings'],
-				]
-			);
-		}
-
-		return $bulk_args;
-	}
-
 	public function bulk_copy_templates( array $args ) {
 		$validate_args = $this->ensure_args( [ 'source', 'from_source', 'from_template_id' ], $args );
 
@@ -1264,9 +1103,13 @@ class Manager {
 			return new \WP_Error( 'template_error', 'Template source not found.' );
 		}
 
-		$bulk_args = 'local' === $args['from_source']
-			? $this->format_args_for_bulk_action_from_local( $args )
-			: $this->format_args_for_bulk_action_from_cloud( $args );
+		$from_source = $this->get_source( $args['from_source'] );
+
+		if ( ! $from_source ) {
+			return new \WP_Error( 'template_error', 'Template source not found.' );
+		}
+
+		$bulk_args = $from_source->format_args_for_bulk_action( $args );
 
 		if ( $source->supports_quota() && ! $this->is_action_to_same_source( $args ) ) {
 			$is_quota_valid = $source->validate_quota( $bulk_args );
